move read_dataset and random_subset functions to new subpackage io
