+++ plantcv/plantcv/io/read_dataset.py
@@ -1,27 +1,11 @@
-# Read dataset of images
-
 import os
 import re
+from plantcv.plantcv import fatal_error
 
-
-def read_dataset(source_path, pattern='', sort=True):
-    """
-    Read a dataset of images as a list of paths.
-
-    Inputs:
-    source_path  = Path to the directory containing the images
-    pattern      = Optional, return only filenames containing the pattern
-    sort         = True by default, sorts the paths alphabetically
-
-    Returns:
-    dataset = List of paths to the images in the source path
-    :param source_path: str
-    :param pattern: str
-    :param sort: bool
-    :return dataset: list
-    """
+def read_dataset(source_path, pattern=''):
     if not os.path.exists(source_path):
         raise IOError("Directory does not exist: {0}".format(source_path))
+
 
     img_path_list = []
     img_extensions = ['.png', '.jpg', '.jpeg', '.tif', '.tiff', '.gif']
@@ -29,13 +13,10 @@ def read_dataset(source_path, pattern='', sort=True):
     for root, dirs, files in os.walk(source_path):
         for file in files:
             # Look for images that contain [pattern] in the name
-            if re.search(pattern, file):
+            if re.search(pattern,file):
                 # Check file type so that only images get selected
                 name, ext = os.path.splitext(file)
                 if ext.lower() in img_extensions:
-                    img_path_list.append(os.path.join(root, file))
-
-    if sort is True:
-        img_path_list = sorted(img_path_list)
+                    img_path_list.append(os.path.join(root,file))
 
     return img_path_list
