+++ plantcv/utils/__init__.py
@@ -1,7 +1,5 @@
-__all__ = ["json2csv", "tabulate_bayes_classes", "sample_images", "read_dataset", "random_subset"]
+__all__ = ["json2csv", "tabulate_bayes_classes", "sample_images"]
 
 from plantcv.utils.converters import json2csv
 from plantcv.utils.converters import tabulate_bayes_classes
 from plantcv.utils.sample_images import sample_images
-from plantcv.utils.read_dataset import read_dataset
-from plantcv.utils.random_subset import random_subset
